requirements.txt
Django==4.2.2
djangorestframework==3.14.0
pinecone-client==2.2.2
python-dotenv==1.0.0
google-cloud-storage==2.10.0
openai==0.27.8
google-cloud-texttospeech==2.14.1
django-cors-headers==4.1.0
dj-database-url==2.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
django-redis==5.3.0